% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getMus}
\alias{getMus}
\title{This function allows to get individual level prognostic estimates.}
\usage{
getMus(fit)
}
\arguments{
\item{fit}{A wsbcf or XBCF fit object.}
}
\value{
An array of prognostic effect point estimates.
}
\description{
This function allows to get individual level prognostic estimates.
}
