% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getTaus}
\alias{getTaus}
\title{This function allows to get individual level treatment estimates.}
\usage{
getTaus(fit)
}
\arguments{
\item{fit}{A wsbcf or XBCF fit object.}
}
\value{
An array of treatment effect point estimates.
}
\description{
This function allows to get individual level treatment estimates.
}
