% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.XBCF.R
\name{predict.XBCF}
\alias{predict.XBCF}
\title{predict.XBCF provides predicted values of a*mu(x_con, pihat) and (b1-b0)*tau(x_mod) from each post-burnin iteration, given input matrices x_con, x_mod and pihat.
The number of columns in matrices should match the number of columns in matrices used for training the model.}
\usage{
\method{predict}{XBCF}(model, x_con, x_mod = x_con, pihat = NULL, burnin = NULL)
}
\arguments{
\item{model}{A trained XBCF model.}

\item{x_con}{An input matrix for the prognostic term of size n by p1. Column order matters: continuos features should all bgo before of categorical.}

\item{x_mod}{An input matrix for the treatment term of size n by p2 (default x_mod = x_con). Column order matters: continuos features should all go before categorical.}

\item{pihat}{An array of propensity score estimates (default is NULL). In the default case propensity scores are estimated using nnet function.}

\item{burnin}{The number of burn-in iterations to discard from prediction (the default value is taken from the trained model).}
}
\value{
A list with two matrices. Each matrix corresponds to a set of draws of predicted values; rows are datapoints, columns are iterations.
}
\description{
predict.XBCF provides predicted values of a*mu(x_con, pihat) and (b1-b0)*tau(x_mod) from each post-burnin iteration, given input matrices x_con, x_mod and pihat.
The number of columns in matrices should match the number of columns in matrices used for training the model.
}
