% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.XBCF.R
\name{predictMuDraws}
\alias{predictMuDraws}
\title{Get post-burnin draws from trained model (prognostic term only)}
\usage{
predictMuDraws(model, x_con, pihat = NULL, burnin = NULL)
}
\arguments{
\item{model}{A trained XBCF model.}

\item{x_con}{An input matrix for the treatment term of size n by p1. Column order matters: continuos features should all go before categorical.}

\item{pihat}{An array of propensity score estimates (default is NULL). In the default case propensity scores are estimated using nnet function.}

\item{burnin}{The number of burn-in iterations to discard from prediction (the default value is taken from the trained model).}
}
\value{
A matrix with a set of draws of predicted prognostic effect estimates; rows are datapoints, columns are iterations.
}
\description{
Get post-burnin draws from trained model (prognostic term only)
}
