% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.XBCF.R
\name{predictTauDraws}
\alias{predictTauDraws}
\title{predictTauDraws provides predicted values of treatment effect (b1-b0)*tau(x_mod) from each post-burnin iteration, given input matrix x_mod.
The number of columns in x_mod here should match the number of columns in x_mod used for training the model.}
\usage{
predictTauDraws(model, x_mod, burnin = NULL)
}
\arguments{
\item{model}{A trained XBCF model.}

\item{x_mod}{An input matrix for the treatment term of size n by p2. Column order matters: continuos features should all go before categorical.}

\item{burnin}{The number of burn-in iterations to discard from prediction (the default value is taken from the trained model).}
}
\value{
A matrix with a set of draws of predicted treatment effect estimates; rows are datapoints, columns are iterations.
}
\description{
predictTauDraws provides predicted values of treatment effect (b1-b0)*tau(x_mod) from each post-burnin iteration, given input matrix x_mod.
The number of columns in x_mod here should match the number of columns in x_mod used for training the model.
}
