\name{XBCF-package}
\alias{XBCF-package}
\docType{package}
\title{
  XBCF: Accelerated Bayesian Causal Forests
}
\description{
This package implements XBCF model for conditional average treatment effect estimation.
}
\details{
This package implements the Accelerated Bayesian Causal Forests approach for conditional average treatment effect estimation. This approach builds on the methodology behind Bayesian Causal Forests outlined in Hahn et al. (2020) and incorporates several improvements to Bayesian Additive Regression Trees implemented by He et al. (2019).
}
\author{
Nikolay Krantsevich, Jingyu He and P. Richard Hahn

Maintainer: Nikolay Krantsevich <krantsevich@asu.edu> and Jingyu He <jingyuhe@cityu.edu.hk>
}
\references{
Krantsevich, Nikolay, Jingyu He, and P. Richard Hahn. "Stochastic Tree Ensembles for Estimating Heterogeneous Effects." URL: https://math. la. asu. edu/~ prhahn/XBCF. pdf (2021).
}
\keyword{ XBCF }
\seealso{
  XBCF
}

