% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XBCF.R
\name{XBCF}
\alias{XBCF}
\title{This function runs the warm-start model on the given data.}
\usage{
XBCF(
  y,
  z,
  x_con,
  x_mod = x_con,
  pihat = NULL,
  num_sweeps = 60,
  burnin = 20,
  max_depth = 50,
  Nmin = 1L,
  num_cutpoints = 100,
  no_split_penality = "Auto",
  mtry_con = 0L,
  mtry_mod = 0L,
  pcat_con = NULL,
  pcat_mod = pcat_con,
  n_trees_con = 30L,
  alpha_con = 0.95,
  beta_con = 1.25,
  tau_con = NULL,
  kap_con = 16,
  s_con = 4,
  pr_scale = FALSE,
  n_trees_mod = 10L,
  alpha_mod = 0.25,
  beta_mod = 3,
  tau_mod = NULL,
  kap_mod = 16,
  s_mod = 4,
  trt_scale = FALSE,
  verbose = FALSE,
  parallel = TRUE,
  random_seed = NULL,
  sample_weights_flag = TRUE,
  a_scaling = TRUE,
  b_scaling = TRUE
)
}
\arguments{
\item{y}{An array of outcome variables of length n (expected to be continuos).}

\item{z}{A binary array of treatment assignments of length n.}

\item{x_con}{An input matrix for the prognostic term of size n by p1. Column order matters: continuos features should all bgo before of categorical.}

\item{x_mod}{An input matrix for the treatment term of size n by p2 (default x_mod = x_con). Column order matters: continuos features should all go beforeof categorical.}

\item{pihat}{An array of propensity score estimates of length n (default is NULL). In the default case propensity scores are evaluated within the wsbcf function with nnet.}

\item{num_sweeps}{The total number of sweeps over data (default is 60).}

\item{max_depth}{The maximum possible depth of a tree.}

\item{Nmin}{The minimum node size.}

\item{num_cutpoints}{The number of adaptive cutpoints considered at each split for continuous variables (default is 100).}

\item{pcat_con}{The number of categorical inputs in the prognostic term input matrix x_con.}

\item{pcat_mod}{The number of categorical inputs in the treatment term input matrix x_mod.}

\item{n_trees_con}{The number of trees in the prognostic forest (default is 30).}

\item{alpha_con}{Base parameter for tree prior on trees in prognostic forest (default is 0.95).}

\item{beta_con}{Power parameter for tree prior on trees in prognostic forest (default is 1.25).}

\item{tau_con}{Prior variance over the mean on on trees in prognostic forest (default is 0.6*var(y)/n_trees_con).}

\item{n_trees_mod}{The number of trees in the treatment forest (default is 10).}

\item{alpha_mod}{Base parameter for tree prior on trees in treatment forest (default is 0.25).}

\item{beta_mod}{Power parameter for tree prior on trees in treatment forest (default is 3).}

\item{tau_mod}{Prior variance over the mean on on trees in treatment forest (default is 0.1*var(y)/n_trees_mod).}

\item{num_burnin}{The number of burn-in sweeps (default is 20).}
}
\value{
A fit file, which contains the draws from the model as well as parameter draws at each sweep.
}
\description{
This function runs the warm-start model on the given data.
}
