% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.XBCF.R
\name{predictTaus}
\alias{predictTaus}
\title{Get point-estimates of treatment effect}
\usage{
predictTaus(model, x_mod, burnin = NULL)
}
\arguments{
\item{model}{A trained XBCF model.}

\item{x_mod}{An input matrix for the treatment term of size n by p2. Column order matters: continuos features should all go before categorical.}

\item{burnin}{The number of burn-in iterations to discard from prediction (the default value is taken from the trained model).}
}
\value{
An array with point-estimates of treatment effect per datapoint in the given matrix.
}
\description{
Get point-estimates of treatment effect
}
